+++ services/david/david.tester.js
@@ -11,6 +11,7 @@ const isDependencyStatus = Joi.string().valid(
 
 t.create('david dependencies (valid)')
   .get('/expressjs/express.json')
+  .timeout(15000)
   .expectBadge({
     label: 'dependencies',
     message: isDependencyStatus,
@@ -18,6 +19,7 @@ t.create('david dependencies (valid)')
 
 t.create('david dev dependencies (valid)')
   .get('/dev/expressjs/express.json')
+  .timeout(15000)
   .expectBadge({
     label: 'dev dependencies',
     message: isDependencyStatus,
@@ -25,6 +27,7 @@ t.create('david dev dependencies (valid)')
 
 t.create('david optional dependencies (valid)')
   .get('/optional/elnounch/byebye.json')
+  .timeout(15000)
   .expectBadge({
     label: 'optional dependencies',
     message: isDependencyStatus,
@@ -32,6 +35,7 @@ t.create('david optional dependencies (valid)')
 
 t.create('david peer dependencies (valid)')
   .get('/peer/webcomponents/generator-element.json')
+  .timeout(15000)
   .expectBadge({
     label: 'peer dependencies',
     message: isDependencyStatus,
@@ -39,6 +43,7 @@ t.create('david peer dependencies (valid)')
 
 t.create('david dependencies with path (valid)')
   .get('/babel/babel.json?path=packages/babel-core')
+  .timeout(15000)
   .expectBadge({
     label: 'dependencies',
     message: isDependencyStatus,
@@ -46,10 +51,12 @@ t.create('david dependencies with path (valid)')
 
 t.create('david dependencies (none)')
   .get('/peer/expressjs/express.json') // express does not specify peer dependencies
+  .timeout(15000)
   .expectBadge({ label: 'peer dependencies', message: 'none' })
 
 t.create('david dependencies (repo not found)')
   .get('/pyvesb/emptyrepo.json')
+  .timeout(15000)
   .expectBadge({
     label: 'dependencies',
     message: 'repo or path not found or david internal error',
@@ -57,6 +64,7 @@ t.create('david dependencies (repo not found)')
 
 t.create('david dependencies (path not found')
   .get('/babel/babel.json?path=invalid/path')
+  .timeout(15000)
   .expectBadge({
     label: 'dependencies',
     message: 'repo or path not found or david internal error',