+++ services/packagist/packagist-version.service.js
@@ -121,8 +121,11 @@ class PackagistVersion extends BasePackagistService {
     }
   }
 
-  async handle({ user, repo }, { include_prereleases, server }) {
-    const includePrereleases = include_prereleases !== undefined
+  async handle(
+    { user, repo },
+    { include_prereleases: includePrereleases, server }
+  ) {
+    includePrereleases = includePrereleases !== undefined
     const json = await this.fetch({
       user,
       repo,