+++ services/symfony/symfony-insight-violations.service.js
@@ -12,7 +12,7 @@ module.exports = class SymfonyInsightViolations extends SymfonyInsightBase {
     {
       title: 'SymfonyInsight Violations',
       namedParams: {
-        projectUuid: '45afb680-d4e6-4e66-93ea-bcfa79eb8a87',
+        projectUuid: '825be328-29f8-44f7-a750-f82818ae9111',
       },
       staticPreview: this.render({
         numViolations: 0,