+++ services/youtube/youtube-base.js
@@ -5,22 +5,34 @@ const { BaseJsonService, NotFound } = require('..')
 const { metric } = require('../text-formatters')
 const { nonNegativeInteger } = require('../validators')
 
+const documentation = `
+<p>By using the YouTube badges provided by Shields.io, you are agreeing to be bound by the YouTube Terms of Service. These can be found here:
+<code>https://www.youtube.com/t/terms</code></p>`
+
 const schema = Joi.object({
-  items: Joi.array()
-    .items(
-      Joi.object({
-        statistics: Joi.object({
+  pageInfo: Joi.object({
+    totalResults: nonNegativeInteger,
+    resultsPerPage: nonNegativeInteger,
+  }).required(),
+  items: Joi.array().items(
+    Joi.object({
+      statistics: Joi.alternatives(
+        Joi.object({
           viewCount: nonNegativeInteger,
           likeCount: nonNegativeInteger,
           dislikeCount: nonNegativeInteger,
           commentCount: nonNegativeInteger,
-        }).required(),
-      })
-    )
-    .required(),
+        }),
+        Joi.object({
+          viewCount: nonNegativeInteger,
+          subscriberCount: nonNegativeInteger,
+        })
+      ),
+    })
+  ),
 }).required()
 
-module.exports = class YouTubeBase extends BaseJsonService {
+class YouTubeBase extends BaseJsonService {
   static category = 'social'
 
   static auth = {
@@ -35,36 +47,49 @@ module.exports = class YouTubeBase extends BaseJsonService {
     namedLogo: 'youtube',
   }
 
-  static renderSingleStat({ statistics, statisticName, videoId }) {
+  static renderSingleStat({ statistics, statisticName, id }) {
     return {
       label: `${statisticName}s`,
       message: metric(statistics[`${statisticName}Count`]),
       style: 'social',
-      link: `https://www.youtube.com/watch?v=${encodeURIComponent(videoId)}`,
+      link: `https://www.youtube.com/${this.type}/${encodeURIComponent(id)}`,
     }
   }
 
-  async fetch({ videoId }) {
+  async fetch({ id }) {
     return this._requestJson(
       this.authHelper.withQueryStringAuth(
         { passKey: 'key' },
         {
           schema,
-          url: 'https://www.googleapis.com/youtube/v3/videos',
+          url: `https://www.googleapis.com/youtube/v3/${this.constructor.type}s`,
           options: {
-            qs: { id: videoId, part: 'statistics' },
+            qs: { id, part: 'statistics' },
           },
         }
       )
     )
   }
 
-  async handle({ videoId }, queryParams) {
-    const json = await this.fetch({ videoId })
-    if (json.items.length === 0) {
-      throw new NotFound({ prettyMessage: 'video not found' })
+  async handle({ channelId, videoId }, queryParams) {
+    const id = channelId || videoId
+    const json = await this.fetch({ id })
+    if (json.pageInfo.totalResults === 0) {
+      throw new NotFound({
+        prettyMessage: `${this.constructor.type} not found`,
+      })
     }
     const statistics = json.items[0].statistics
-    return this.constructor.render({ statistics, videoId }, queryParams)
+    return this.constructor.render({ statistics, id }, queryParams)
   }
 }
+
+class YouTubeVideoBase extends YouTubeBase {
+  static type = 'video'
+}
+
+class YouTubeChannelBase extends YouTubeBase {
+  static type = 'channel'
+}
+
+module.exports = { documentation, YouTubeVideoBase, YouTubeChannelBase }